refactor: 구조 변경 interface의 common 디렉토리를 상위 디렉토리로 이동했습니다
