+++ src/main/java/hhplus/hhplusconcertreservation/interfaces/presetation/exception/ExceptionController.java
@@ -1,4 +1,4 @@
-package hhplus.hhplusconcertreservation.interfaces.common.controller;
+package hhplus.hhplusconcertreservation.interfaces.presetation.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,7 +7,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 import hhplus.hhplusconcertreservation.domain.common.exception.CoreException;
-import hhplus.hhplusconcertreservation.interfaces.common.dto.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
